+++ src/main/java/de/gematik/tim/test/glue/api/fhir/organisation/endpoint/UseEndpointAbility.java
@@ -16,6 +16,12 @@
 
 package de.gematik.tim.test.glue.api.fhir.organisation.endpoint;
 
+import static de.gematik.tim.test.glue.api.TestdriverApiPath.ENDPOINT_ID_VARIABLE;
+import static de.gematik.tim.test.glue.api.fhir.organisation.endpoint.DeleteEndpointTask.deleteEndPoint;
+import static de.gematik.tim.test.glue.api.fhir.organisation.endpoint.UseEndpointAbility.EndpointInfo;
+import static java.util.Objects.isNull;
+import static java.util.Objects.requireNonNull;
+
 import de.gematik.tim.test.glue.api.MultiTargetAbility;
 import de.gematik.tim.test.glue.api.TestdriverApiAbility;
 import de.gematik.tim.test.glue.api.fhir.organisation.healthcareservice.UseHealthcareServiceAbility;
@@ -23,17 +29,11 @@ import io.restassured.specification.RequestSpecification;
 import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.Task;
 
-import static de.gematik.tim.test.glue.api.TestdriverApiPath.ENDPOINT_ID_VARIABLE;
-import static de.gematik.tim.test.glue.api.fhir.organisation.endpoint.DeleteEndpointTask.deleteEndPoint;
-import static de.gematik.tim.test.glue.api.fhir.organisation.endpoint.UseEndpointAbility.EndpointInfo;
-import static java.util.Objects.isNull;
-import static java.util.Objects.requireNonNull;
-
 public class UseEndpointAbility extends MultiTargetAbility<String, EndpointInfo> implements
     TestdriverApiAbility {
 
   public static <T extends Actor> void addEndpointToActorForHS(T actor, String endpointName,
-                                                               String endpointId, String hsName) {
+      String endpointId, String hsName) {
     UseEndpointAbility ability = actor.abilityTo(UseEndpointAbility.class);
     if (isNull(ability)) {
       actor.can(ability = new UseEndpointAbility());
@@ -65,5 +65,6 @@ public class UseEndpointAbility extends MultiTargetAbility<String, EndpointInfo>
   }
 
   protected record EndpointInfo(String endpointId, String hsName) {
+
   }
 }