+++ src/main/java/de/gematik/tim/test/glue/api/fhir/organisation/location/FhirGetLocationListQuestion.java
@@ -17,12 +17,15 @@
 package de.gematik.tim.test.glue.api.fhir.organisation.location;
 
 import static de.gematik.tim.test.glue.api.TestdriverApiEndpoint.GET_LOCATIONS;
+import static de.gematik.tim.test.glue.api.utils.GlueUtils.getMapper;
+import static de.gematik.tim.test.glue.api.utils.GlueUtils.getResourcesFromSearchResult;
+import static de.gematik.tim.test.models.FhirResourceTypeDTO.LOCATION;
 import static java.util.Objects.requireNonNull;
 import static net.serenitybdd.rest.SerenityRest.lastResponse;
 
 import de.gematik.tim.test.glue.api.fhir.organisation.healthcareservice.HealthcareSpecificTask;
 import de.gematik.tim.test.models.FhirLocationDTO;
-import de.gematik.tim.test.models.FhirLocationsDTO;
+import de.gematik.tim.test.models.FhirSearchResultDTO;
 import java.util.List;
 import java.util.Optional;
 import net.serenitybdd.screenplay.Actor;
@@ -31,32 +34,35 @@ import net.serenitybdd.screenplay.Question;
 public class FhirGetLocationListQuestion extends HealthcareSpecificTask implements
     Question<List<FhirLocationDTO>> {
 
-  private String locationNameFilter;
+  private String locationName;
 
   public static FhirGetLocationListQuestion getLocationList() {
     return new FhirGetLocationListQuestion();
   }
 
   public FhirGetLocationListQuestion filterForName(String locationName) {
-    this.locationNameFilter = locationName;
+    this.locationName = locationName;
     return this;
   }
 
   @Override
   public List<FhirLocationDTO> answeredBy(Actor actor) {
     super.performAs(actor);
-    Optional<String> locationId = Optional.ofNullable(locationNameFilter)
+    Optional<String> locationId = Optional.ofNullable(locationName)
         .map(name -> actor.abilityTo(UseLocationAbility.class).getTarget(name).locationId());
 
     actor.attemptsTo(GET_LOCATIONS.request());
 
-    List<FhirLocationDTO> fhirLocationDTOs = requireNonNull(
-        lastResponse().body().as(FhirLocationsDTO.class).getLocations());
+    FhirSearchResultDTO result = requireNonNull(lastResponse().body().as(FhirSearchResultDTO.class, getMapper()));
+
+    List<FhirLocationDTO> fhirLocationDTOs = getResourcesFromSearchResult(result, LOCATION, FhirLocationDTO.class);
+
     return fhirLocationDTOs.stream()
         .filter(location -> locationId
-            .map(id -> id.equals(location.getLocationId()))
+            .map(id -> id.equals(location.getId()))
             .orElse(true))
         .toList();
   }
+
 
 }