+++ src/main/java/de/gematik/tim/test/glue/api/fhir/organisation/location/UseLocationAbility.java
@@ -16,18 +16,18 @@
 
 package de.gematik.tim.test.glue.api.fhir.organisation.location;
 
+import static de.gematik.tim.test.glue.api.TestdriverApiPath.LOCATION_ID_VARIABLE;
+import static de.gematik.tim.test.glue.api.fhir.organisation.location.DeleteLocationTask.deleteLocation;
+import static de.gematik.tim.test.glue.api.fhir.organisation.location.UseLocationAbility.LocationInfo;
+import static java.util.Objects.isNull;
+import static java.util.Objects.requireNonNull;
+
 import de.gematik.tim.test.glue.api.MultiTargetAbility;
 import de.gematik.tim.test.glue.api.TestdriverApiAbility;
 import de.gematik.tim.test.glue.api.fhir.organisation.healthcareservice.UseHealthcareServiceAbility;
 import io.restassured.specification.RequestSpecification;
 import net.serenitybdd.screenplay.Actor;
 import net.serenitybdd.screenplay.Task;
-
-import static de.gematik.tim.test.glue.api.TestdriverApiPath.LOCATION_ID_VARIABLE;
-import static de.gematik.tim.test.glue.api.fhir.organisation.location.DeleteLocationTask.deleteLocation;
-import static de.gematik.tim.test.glue.api.fhir.organisation.location.UseLocationAbility.LocationInfo;
-import static java.util.Objects.isNull;
-import static java.util.Objects.requireNonNull;
 
 public class UseLocationAbility extends MultiTargetAbility<String, LocationInfo> implements
     TestdriverApiAbility {
@@ -64,6 +64,7 @@ public class UseLocationAbility extends MultiTargetAbility<String, LocationInfo>
   }
 
   protected record LocationInfo(String locationId, String hsName) {
+
   }
 }